create tab and project list within components
