+++ src/components/Tabs/index.tsx
@@ -1,0 +1,3 @@
+export * from './TabContainer';
+export * from './TabList';
+export * from './TabPanelContentContainer';
